requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.1.0
Babel==2.9.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.18
django-cors-headers==4.6.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==1.10.21
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
yarl==1.18.3

djangorestframework~=3.15.2